refactor(host-application): move runtime startup ownership into runtimes
