+++ apps/desktop/src-tauri/crates/host-application/src/app_service/opencode_runtime.rs
@@ -12,11 +12,7 @@ pub(crate) use process_lifecycle::{
     resolve_opencode_binary_path, terminate_child_process, terminate_process_by_pid,
     wait_for_process_exit_by_pid,
 };
-pub use startup_readiness::OpencodeStartupWaitFailure;
-pub(crate) use startup_readiness::{
-    wait_for_local_server_with_process, OpencodeStartupReadinessPolicy, OpencodeStartupWaitReport,
-    StartupCancelEpoch,
-};
+pub(crate) use startup_readiness::{wait_for_local_server_with_process, StartupCancelEpoch};
 
 impl AppService {
     pub(crate) fn spawn_opencode_server(
