+++ apps/desktop/src-tauri/crates/host-application/src/app_service/opencode_runtime/startup_readiness/mod.rs
@@ -2,10 +2,7 @@ mod policy;
 mod probe_runtime;
 mod wait_loop;
 
-pub use policy::OpencodeStartupWaitFailure;
-pub(crate) use policy::{
-    OpencodeStartupReadinessPolicy, OpencodeStartupWaitReport, StartupCancelEpoch,
-};
+pub(crate) use policy::StartupCancelEpoch;
 pub(crate) use wait_loop::wait_for_local_server_with_process;
 
 #[cfg(test)]
