+++ apps/desktop/src-tauri/crates/host-application/src/app_service/runtime_orchestrator/repo_health/mcp.rs
@@ -8,7 +8,7 @@ impl AppService {
             .chain()
             .find_map(|cause| {
                 cause
-                    .downcast_ref::<OpencodeStartupWaitFailure>()
+                    .downcast_ref::<RuntimeStartupWaitFailure>()
                     .map(|failure| {
                         if failure.reason == "timeout" {
                             RepoRuntimeStartupFailureKind::Timeout
@@ -30,7 +30,7 @@ impl AppService {
     ) -> Option<String> {
         error.chain().find_map(|cause| {
             cause
-                .downcast_ref::<OpencodeStartupWaitFailure>()
+                .downcast_ref::<RuntimeStartupWaitFailure>()
                 .map(|failure| failure.reason.to_string())
         })
     }
