+++ apps/desktop/src-tauri/crates/host-application/src/lib.rs
@@ -4,8 +4,8 @@ pub use app_service::build_orchestrator::{BuildResponseAction, CleanupMode};
 pub use app_service::{
     AppService, DevServerEmitter, HookTrustConfirmationPort, HookTrustConfirmationRequest,
     OdtCreateTaskInput, OdtHostBridgeReady, OdtSearchTasksInput, OdtSearchTasksResult,
-    OdtSetPlanResult, OdtSetPullRequestResult, OdtSetSpecResult, OdtTaskDocumentsRead,
-    OdtTaskResult, OdtTaskSummary, OpencodeStartupWaitFailure, PreparedHookTrustChallenge,
+    OdtSetPlanResult, OdtSetPullRequestResult, OdtSetSpecResult, OdtTaskDocumentsRead, OdtTaskResult,
+    OdtTaskSummary, PreparedHookTrustChallenge, RuntimeStartupWaitFailure,
     RepoConfigUpdate, RepoPullRequestSyncResult, RepoSettingsUpdate, RunEmitter,
     WorkspaceSettingsSnapshotUpdate,
 };
