+++ apps/desktop/src-tauri/src/lib.rs
@@ -200,7 +200,7 @@ pub(crate) fn as_error<T>(result: anyhow::Result<T>) -> Result<T, String> {
 pub(crate) fn runtime_ensure_failure_kind(error: &anyhow::Error) -> Option<&'static str> {
     error.chain().find_map(|cause| {
         cause
-            .downcast_ref::<host_application::OpencodeStartupWaitFailure>()
+            .downcast_ref::<host_application::RuntimeStartupWaitFailure>()
             .map(|failure| match failure.reason {
                 "timeout" => "timeout",
                 _ => "error",
